requirements.txt
openai
faiss-cpu
numpy
rank-bm25
psycopg2-binary
python-dotenv
tiktoken
nltk
PyMuPDF
pytest
python-telegram-bot